@jngadiub 2019-05-09
Does the difference come from the fact that you do not subtract the mean anymore as in 

https://github.com/hls-fpga-machine-learning/hls4ml/blob/master/nnet_utils/nnet_batchnorm.h#L94

I do not understand though why the nnet_batchnorm.h is not changed accordingly in your commit.